requirements.txt
# LiveKit Agents Framework
livekit-agents>=1.0.0
livekit-plugins-sarvam>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-silero>=1.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Redis for state management
redis>=5.0.0

# FastAPI for API & Web Server
fastapi>=0.109.0
uvicorn>=0.27.0

# Static file serving
aiofiles>=23.0.0

# Utilities
loguru>=0.7.0
tenacity>=8.2.0

# LiveKit Server SDK (for token generation)
livekit-api>=0.6.0
